% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_values.R
\docType{data}
\name{critical_values}
\alias{critical_values}
\title{Critical values for the CVR value at level of 0.05}
\format{
A data frame with 39 rows and 5 variables:
\describe{
\item{N}{Number of experts}
\item{Exact}{Exact binomial Ayre and Scally, 2014.}
\item{Chi}{Chi square by Tristán-López, 2008.}
\item{Lawshe}{Lawshe 1975.}
\item{Bayes}{Bayesian as suggestd by Baghestani et al., 2019.}
}
}
\source{
\url{10.3389/feduc.2023.1271335}
}
\usage{
critical_values
}
\description{
Jeldres et al (2023) compilation of various critical values for
the CVR value at level of 0.05. Full source is available at:
Romero Jeldres, M., Díaz Costa, E., & Faouzi Nadim, T. (2023).
A review of Lawshe’s method for calculating content validity in
the social sciences. In Frontiers in Education (Vol. 8, p. 1271335).
https://doi.org/10.3389/feduc.2023.1271335
}
\keyword{datasets}
