% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI-package.R
\docType{package}
\name{CVI-package}
\alias{CVI}
\alias{CVI-package}
\title{CVI: Calculates Content Validity Indices}
\description{
Offers tools for calculating various content validity indices (item level, ratio, kappa adjusted) and validity on a scale level.
}
\author{
\strong{Maintainer}: Tuomas Eerola \email{tuomas.eerola@durham.ac.uk} (\href{https://orcid.org/0000-0002-2896-929X}{ORCID})

}
\keyword{internal}
