% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_check.R
\name{CVI_check}
\alias{CVI_check}
\title{Check assumptions if the data is in correct format}
\usage{
CVI_check(data = NULL, response_options = "default")
}
\arguments{
\item{data}{data frame to be processed, where the minimal requirements are:
ID, name, value}

\item{response_options}{relate to ratings of the items
where the default is 'Extremely Relevant','Moderately Relevant','Slightly Relevant','Not Relevant'}
}
\value{
data frame with the same structure as the input
}
\description{
CVI package assumes that the data is in tidy format
}
