% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_decision.R
\name{CVI_decision}
\alias{CVI_decision}
\title{Return decision based on CVI}
\usage{
CVI_decision(CVI = NULL)
}
\arguments{
\item{CVI}{value to be processed}
}
\value{
data frame with the same structure as the input
}
\description{
Derived from Abdollahpour et al. (2010)
}
