% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_item.R
\name{CVI_item}
\alias{CVI_item}
\title{Run content validity indices for items}
\usage{
CVI_item(data = NULL)
}
\arguments{
\item{data}{data frame to be processed, where the minimal requirements are:
ID, name, value}
}
\value{
data frame with the same structure as the input
}
\description{
Assumes that the ratings are on a specific scale of 1-4
Calculates CVI for items
}
