% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_scale.R
\name{CVI_scale}
\alias{CVI_scale}
\title{Returns overall CVI for the scale using the item information}
\usage{
CVI_scale(items = NULL, method = "average")
}
\arguments{
\item{items}{items as data frame}

\item{method}{relate to type of summary, either
average or universal
Source:
Polit DF, Beck CT, Owen SV. Is the CVI an acceptable indicator of content validity? Appraisal and recommendations. Res Nurs Health. 2007;30(4):459–67. doi: 10.1002/nur.20199.}
}
\value{
data frame with the same structure as the input
}
\description{
Assumes items have been calculated previously
}
