% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVI_visualise.R
\name{CVI_visualise}
\alias{CVI_visualise}
\title{Creates a plot of the CVI per items}
\usage{
CVI_visualise(
  data,
  yvar = "CVI.I.adj",
  colorvar = "KappaFit",
  showlegend = FALSE
)
}
\arguments{
\item{data}{items as data frame}

\item{yvar}{which variable to plot on the Y-axis}

\item{colorvar}{which variable to plot as the colour}

\item{showlegend}{option to show legend (false is the default)}
}
\value{
data frame with the same structure as the input
}
\description{
Assumes item CVIs have been calculated previously
}
