% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVR_decision.R
\name{CVR_decision}
\alias{CVR_decision}
\title{Return decision based on CVR}
\usage{
CVR_decision(CVR = NULL, N_experts = NULL, Method = "Exact")
}
\arguments{
\item{CVR}{CVR value}

\item{N_experts}{Number of experts}

\item{Method}{Method for calculating the decision. Default is "Exact"
but 'Bayes', 'Chi' and 'Latche' are also possible.}
}
\value{
data frame with the same structure as the input
}
\description{
Using the critical values from Jeldres et al 2023
(see critical_values dataset).
Source:
Romero Jeldres, M., Díaz Costa, E., & Faouzi Nadim, T. (2023).
A review of Lawshe’s method for calculating content validity in
the social sciences. In Frontiers in Education (Vol. 8, p. 1271335).
https://doi.org/10.3389/feduc.2023.1271335
}
