% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{ELBO}
\alias{ELBO}
\title{Title Expected lower bound (ELBO)}
\usage{
ELBO(N, D, T0, s1, s2, L20, X, W1, W2, L1, L2, Plog)
}
\arguments{
\item{N}{number of observed data}

\item{D}{dimension of each observed data}

\item{T0}{total clusters fixed for the variational distribution}

\item{s1}{shape parameter for the prior distribution of alpha}

\item{s2}{rate parameter for the prior distribution of alpha}

\item{L20}{precision parameter for the prior distribution of eta_i's}

\item{X}{the observed data, N x D matrix}

\item{W1}{shape parameter for the posterior distribution of alpha}

\item{W2}{rate parameter for the posterior distribution of alpha}

\item{L1}{1st parameter for the posterior distribution of eta_i's}

\item{L2}{Precision parameter for the posterior distribution of eta_i's}

\item{Plog}{Latent probability values}
}
\value{
value of the ELBO function
}
\description{
Title Expected lower bound (ELBO)
}
\examples{
ELBO(N = 100, D = 2, T0 = 10, s1 = 0.01, s2 = 0.01, L20 = 0.01,
X = matrix(c(rep(0, 50), rep(10, 50), rep(0, 50), rep(10, 50)), nrow = 100,
ncol = 2), W1 = 0.01, W2 = 0.01, L1 = matrix(0.01, nrow = 10, ncol = 2),
L2 = matrix(0.01, nrow = 10, ncol = 1),
Plog = matrix(-3, nrow = 100, ncol = 10))
}
