% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO_sigma.R
\name{ELBO_sigma}
\alias{ELBO_sigma}
\title{Title Expected lower bound (ELBO) for varying sigma^2}
\usage{
ELBO_sigma(N, D, T0, s1, s2, L20, b1, b2, X, W1, W2, L1, L2, G1, G2, Plog)
}
\arguments{
\item{N}{number of observed data}

\item{D}{dimension of each observed data}

\item{T0}{total clusters fixed for the variational distribution}

\item{s1}{shape parameter for the prior distribution of alpha}

\item{s2}{rate parameter for the prior distribution of alpha}

\item{L20}{precision parameter for the prior distribution of eta_i's}

\item{b1}{shape parameter for prior distribution of sigma^2}

\item{b2}{rate parameter for prior distribution of sigma^2}

\item{X}{the observed data, N x D matrix}

\item{W1}{shape parameter for the posterior distribution of alpha}

\item{W2}{rate parameter for the posterior distribution of alpha}

\item{L1}{1st parameter for the posterior distribution of eta_i's}

\item{L2}{Precision parameter for the posterior distribution of eta_i's}

\item{G1}{shape parameter for the posterior distribution of sigma^2}

\item{G2}{rate parameter for the posterior distribution of sigma^2}

\item{Plog}{Latent probability values}
}
\value{
value of the ELBO function
}
\description{
Title Expected lower bound (ELBO) for varying sigma^2
}
\examples{
ELBO_sigma(N = 100, D = 2, T0 = 10, s1 = 0.01, s2 = 0.01,
             L20 = 0.01, b1 = 0.01, b2 = 0.01,
             X = matrix(c(rep(0, 50), rep(10, 50), rep(0, 50), rep(10, 50)),
                 nrow = 100, ncol = 2),
             W1 = 0.01, W2 = 0.01, L1 = matrix(0.01, nrow = 10, ncol = 2),
             L2 = matrix(0.01, nrow = 10, ncol = 1), G1 = 0.01, G2 = 0.01,
             Plog = matrix(-3, nrow = 100, ncol = 10))
}
