% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM.R
\docType{data}
\name{FM}
\alias{FM}
\title{FM}
\format{
A data tibble with 255 rows and 3 variables:
\describe{
\item{ID}{Expert ID}
\item{name}{Item name}
\item{value}{Rating value}
}
}
\usage{
FM
}
\description{
Rating data from Connor et al., (in prep) containing 15 experts' ratings
of relevance for Focus-Motivation construct as measured with 17 items
that relate to emotional episodes involved in music. The ratings were done
on a 4-point scale (1 = Not Relevant, 2 = Slightly Relevant, 3 = Moderately Relevant,
4 = Extremely Relevant).
}
\keyword{datasets}
