% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_chance_adjusted.R
\name{kappa_chance_adjusted}
\alias{kappa_chance_adjusted}
\title{Returns kappa adjusted for chance agreement}
\usage{
kappa_chance_adjusted(N = NULL, A = NULL, CVI = NULL)
}
\arguments{
\item{N}{Number of experts}

\item{A}{Amount of experts agreed}

\item{CVI}{CVI value to be adjusted}
}
\value{
data frame with the same structure as the input
}
\description{
Assumes that item CVIs have been calculated previously
}
