% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_decision.R
\name{kappa_decision}
\alias{kappa_decision}
\title{Return decision based on kappa}
\usage{
kappa_decision(CVI = NULL)
}
\arguments{
\item{CVI}{value to be processed}
}
\value{
data frame with the same structure as the input
}
\description{
Derived from Cicchetti, D. V., & Sparrow, S. A. (1981)
}
