% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CVIoutput.R
\name{plot.CVIoutput}
\alias{plot.CVIoutput}
\title{S3 plotting function for \code{CVIoutput}objects'}
\usage{
\method{plot}{CVIoutput}(x, y = NULL, ...)
}
\arguments{
\item{x}{a CVIoutput object}
}
\description{
S3 plotting function for \code{CVIoutput}objects'
}
