% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM.R
\docType{data}
\name{FM}
\alias{FM}
\title{FM}
\format{
A data tibble with 300 rows and 3 variables:
\describe{
\item{ID}{Expert ID}
\item{name}{Item name}
\item{value}{Rating value}
}
}
\usage{
FM
}
\description{
Rating data example from Connor et al., (in prep)
This data relates to the Focus-Motivation construct related to
music-related emotional experiences.
}
\keyword{datasets}
